# synkronus/openapi/synkronus.yaml
openapi: 3.0.0
info:
  title: Synkronus API
  version: 1.0.5
paths:
...
  /api/app-bundle/changes:
    get:
      operationId: getAppBundleChanges
      summary: Get changes between two app bundle versions
      description: Compares two versions of the app bundle and returns detailed changes
      security:
        - bearerAuth: [read-only, read-write]
      parameters:
        - name: current
          in: query
          required: false
          schema:
            type: string
          description: The current version (defaults to latest)
        - name: target
          in: query
          required: false
          schema:
            type: string
          description: The target version to compare against (defaults to previous version)
        - name: x-ode-version
          in: header
          required: true
          schema:
            type: string
            pattern: '^\d+\.\d+\.\d+$'
            example: '1.0.0'
          description: Required client version header using semantic versioning (MAJOR.MINOR.PATCH). Major version must be compatible with the server.
      responses:
        '200':
          description: Successfully retrieved changes between versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeLog'
        '400':
          description: Invalid version format or parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: One or both versions not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/app-bundle/manifest:
    get:
      operationId: getAppBundleManifest
      summary: Get the current custom app bundle manifest
      security:
        - bearerAuth: [read-only, read-write]
      parameters:
        - name: x-ode-version
          in: header
          required: true
          schema:
            type: string
            pattern: '^\d+\.\d+\.\d+$'
            example: '1.0.0'
          description: Required client version header using semantic versioning (MAJOR.MINOR.PATCH). Major version must be compatible with the server.
        - name: x-ode-client-id
          in: header
          required: false
          schema:
            type: string
          description: Optional client instance id for correlating app bundle checks with presence.
      responses:
        '200':
          description: Bundle file list
          headers:
            etag:
              schema:
                type: string
              description: Hash of the manifest for caching
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppBundleManifest'
  /api/app-bundle/download/{path}:
    get:
      operationId: downloadAppBundleFile
      summary: Download a specific file from the app bundle
      security:
        - bearerAuth: [read-only, read-write]
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
        - name: preview
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: If true, returns the file from the latest version including unreleased changes
        - name: if-none-match
          in: header
          schema:
            type: string
        - name: x-ode-version
          in: header
          required: true
          schema:
            type: string
            pattern: '^\d+\.\d+\.\d+$'
            example: '1.0.0'
          description: Required client version header using semantic versioning (MAJOR.MINOR.PATCH). Major version must be compatible with the server.
      responses:
        '200':
          description: File content
          headers:
            etag:
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '304':
          description: Not Modified
  /api/app-bundle/versions:
    get:
      operationId: getAppBundleVersions
      summary: Get a list of available app bundle versions
      security:
        - bearerAuth: [read-only, read-write]
      parameters:
        - name: x-ode-version
          in: header
          required: true
          schema:
            type: string
            pattern: '^\d+\.\d+\.\d+$'
            example: '1.0.0'
          description: Required client version header using semantic versioning (MAJOR.MINOR.PATCH). Major version must be compatible with the server.
      responses:
        '200':
          description: List of available app bundle versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppBundleVersions'
  /api/app-bundle/push:
    post:
      operationId: pushAppBundle
      summary: Upload a new app bundle (admin only)
      security:
        - bearerAuth: [admin]
      parameters:
        - name: x-ode-version
          in: header
          required: true
          schema:
            type: string
            pattern: '^\d+\.\d+\.\d+$'
            example: '1.0.0'
          description: Required client version header using semantic versioning (MAJOR.MINOR.PATCH). Major version must be compatible with the server.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                bundle:
                  type: string
                  format: binary
                  description: ZIP file containing the new app bundle
      responses:
        '200':
          description: App bundle successfully uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppBundlePushResponse'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden - Admin role required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '413':
          description: File too large
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/app-bundle/switch/{version}:
    post:
      operationId: switchAppBundleVersion
      summary: Switch to a specific app bundle version (admin only)
      security:
        - bearerAuth: [admin]
      parameters:
        - name: version
          in: path
          required: true
          schema:
            type: string
          description: Version identifier to switch to
        - name: x-ode-version
          in: header
          required: true
          schema:
            type: string
            pattern: '^\d+\.\d+\.\d+$'
            example: '1.0.0'
          description: Required client version header using semantic versioning (MAJOR.MINOR.PATCH). Major version must be compatible with the server.
      responses:
        '200':
          description: Successfully switched to the specified version
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Switched to app bundle version 20250507-123456"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden - Admin role required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Version not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/auth/login:
    post:
      operationId: login
      summary: Authenticate user and return JWT tokens
      description: Obtain a JWT token by providing username and password
      parameters:
        - name: x-ode-version
          in: header
          required: true
          schema:
            type: string
            pattern: '^\d+\.\d+\.\d+$'
            example: '1.0.0'
          description: Required client version header using semantic versioning (MAJOR.MINOR.PATCH). Major version must be compatible with the server.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  description: User's username
                password:
                  type: string
                  format: password
                  description: User's password
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Authentication failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh JWT token
      description: Obtain a new JWT token using a refresh token
      parameters:
        - name: x-ode-version
          in: header
          required: true
          schema:
            type: string
            pattern: '^\d+\.\d+\.\d+$'
            example: '1.0.0'
          description: Required client version header using semantic versioning (MAJOR.MINOR.PATCH). Major version must be compatible with the server.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  description: Refresh token obtained from login or previous refresh
      responses:
        '200':
          description: Token refresh successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Invalid or expired refresh token
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/users/create:
    post:
      operationId: createUser
      summary: Create a new user (admin only)
      description: Create a new user with specified username, password, and role
      security:
        - bearerAuth: [admin]
      parameters:
        - name: x-ode-version
          in: header
          required: true
          schema:
            type: string
            pattern: '^\d+\.\d+\.\d+$'
            example: '1.0.0'
          description: Required client version header using semantic versioning (MAJOR.MINOR.PATCH). Major version must be compatible with the server.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, role]
              properties:
                username:
                  type: string
                  description: New user's username
                password:
                  type: string
                  format: password
                  description: New user's password
                role:
                  type: string
                  enum: [read-only, read-write, admin]
                  description: User's role
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden - Admin role required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Conflict - Username already exists
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/users:
    get:
      operationId: listUsers
      summary: List all users (admin only)
      description: |
        Retrieve a list of all users in the system. Admin access required.
        Each item may include optional `presence` (last-seen per client, bundle/Ode hints) when the server has recorded activity.
      security:
        - bearerAuth: [admin]
      parameters:
        - name: x-ode-version
          in: header
          required: true
          schema:
            type: string
            pattern: '^\d+\.\d+\.\d+$'
            example: '1.0.0'
          description: Required client version header using semantic versioning (MAJOR.MINOR.PATCH). Major version must be compatible with the server.
        - name: x-ode-client-id
          in: header
          required: false
          schema:
            type: string
          description: Optional client instance id (browser/CLI); used for presence when sent with authenticated requests.
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserListItem'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden - Admin role required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /api/users/reset-password:
    post:
      operationId: resetUserPassword
      summary: Reset user password (admin only)
      description: Reset password for a specified user
      security:
        - bearerAuth: [admin]
      parameters:
        - name: x-ode-version
          in: header
          required: true
          schema:
            type: string
            pattern: '^\d+\.\d+\.\d+$'
            example: '1.0.0'
          description: Required client version header using semantic versioning (MAJOR.MINOR.PATCH). Major version must be compatible with the server.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, newPassword]
              properties:
                username:
                  type: string
                  description: Username of the user whose password is being reset
                newPassword:
                  type: string
                  format: password
                  description: New password for the user
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password reset successfully"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden - Admin role required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/users/change-password:
    post:
      operationId: changePassword
      summary: Change user password (authenticated user)'s password
      description: Change password for the currently authenticated user
      security:
        - bearerAuth: [read-only, read-write, admin]
      parameters:
        - name: x-ode-version
          in: header
          required: true
          schema:
            type: string
            pattern: '^\d+\.\d+\.\d+$'
            example: '1.0.0'
          description: Required client version header using semantic versioning (MAJOR.MINOR.PATCH). Major version must be compatible with the server.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                  format: password
                  description: Current password for verification
                newPassword:
                  type: string
                  format: password
                  description: New password to set
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password changed successfully"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized or incorrect current password
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/sync/pull:
    post:
      operationId: syncPull
      summary: Pull updated records since last sync
      description: |
        Retrieves records that have changed since a specified version.
        
        **Pagination Pattern:**
        1. Send initial request with `since.version` (or omit for all records)
        2. Process returned records
        3. If `has_more` is true, make next request using `change_cutoff` as the new `since.version`
        4. Repeat until `has_more` is false
        
        Example pagination flow:
        - Request 1: `since: {version: 100}` → Response: `change_cutoff: 150, has_more: true`
        - Request 2: `since: {version: 150}` → Response: `change_cutoff: 200, has_more: false`
      security:
        - bearerAuth: [read-only, read-write]
      parameters:
        - name: schemaType
          in: query
          schema:
            type: string
          description: Filter by schemaType
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
          description: Maximum number of records to return

        - name: x-ode-version
          in: header
          required: true
          schema:
            type: string
            pattern: '^\d+\.\d+\.\d+$'
            example: '1.0.0'
          description: Required client version header using semantic versioning (MAJOR.MINOR.PATCH). Major version must be compatible with the server.
        - name: x-ode-client-id
          in: header
          required: false
          schema:
            type: string
          description: Optional client instance id; improves per-device presence when combined with sync body `client_id`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncPullRequest'
      responses:
        '200':
          description: Sync data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncPullResponse'
  /api/sync/push:
    post:
      operationId: syncPush
      summary: Push new or updated records to the server
      security:
        - bearerAuth: [read-write]
      parameters:
        - name: x-ode-version
          in: header
          required: true
          schema:
            type: string
            pattern: '^\d+\.\d+\.\d+$'
            example: '1.0.0'
          description: Required client version header using semantic versioning (MAJOR.MINOR.PATCH). Major version must be compatible with the server.
        - name: x-ode-client-id
          in: header
          required: false
          schema:
            type: string
          description: Optional client instance id; improves per-device presence when combined with sync body `client_id`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncPushRequest'
      responses:
        '200':
          description: Sync result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncPushResponse'
...
  /api/dataexport/parquet:
    get:
      summary: Download a ZIP archive of Parquet exports
      description: >
        Returns a ZIP file containing multiple Parquet files,
        each representing a flattened export of observations per form type.
        Supports downloading the entire dataset as separate Parquet files bundled together.
      operationId: getParquetExportZip
      tags:
        - DataExport
      responses:
        '200':
          description: ZIP archive stream containing Parquet files
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - bearerAuth: [read-only, read-write, admin]
  /api/dataexport/raw-json:
    get:
      operationId: getRawJsonExportZip
      summary: Download a ZIP archive of per-observation JSON files
      description: >
        Returns a ZIP archive where each non-deleted observation is one JSON file,
        grouped by form type folder. Each file contains metadata fields and a nested `data` object with the form payload.
      tags:
        - DataExport
      responses:
        '200':
          description: ZIP archive stream containing JSON files
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - bearerAuth: [read-only, read-write, admin]
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Failed to get version information"
    ChangeLog:
      type: object
      properties:
        compare_version_a:
          type: string
        compare_version_b:
          type: string
        form_changes:
          type: boolean
        ui_changes:
          type: boolean
        new_forms:
          type: array
          items:
            $ref: '#/components/schemas/FormDiff'
        removed_forms:
          type: array
          items:
            $ref: '#/components/schemas/FormDiff'
        modified_forms:
          type: array
          items:
            $ref: '#/components/schemas/FormModification'
...
    AppBundleManifest:
      type: object
      required: [files, version, generatedAt, hash]
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/AppBundleFile'
        version:
          type: string
        generatedAt:
          type: string
          format: date-time
        hash:
          type: string
...
    AppBundleVersions:
      type: object
      required: [versions]
      properties:
        versions:
          type: array
          items:
            type: string
...
    AppBundlePushResponse:
      type: object
      required: [message, manifest]
      properties:
        message:
          type: string
        manifest:
          $ref: '#/components/schemas/AppBundleManifest'
    AuthResponse:
      type: object
      required: [token, refreshToken, expiresAt]
      properties:
        token:
          type: string
        refreshToken:
          type: string
        expiresAt:
          type: integer
          format: int64
    UserResponse:    
      type: object
      required: [username, role, createdAt]
      properties:
        username:
          type: string
        role:
          type: string
          enum: [read-only, read-write, admin]
        createdAt:
          type: string
          format: date-time
    UserListItem:
      type: object
      required: [id, username, role, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
          enum: [read-only, read-write, admin]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        presence:
          $ref: '#/components/schemas/UserPresenceSummary'
...
    SyncPullRequest:
      type: object
      required: [client_id]
      properties:
        client_id:
          type: string
        since:
          type: object
          description: Optional pagination cursor indicating the last seen change
          properties:
            version:
              type: integer
            id:
              type: string
        schema_types:
          type: array
          items:
            type: string
...
    SyncPushRequest:
      type: object
      required: [transmission_id, client_id, records]
      properties:
        transmission_id:
          type: string
          format: uuid
        client_id:
          type: string
        records:
          type: array
          items:
            $ref: '#/components/schemas/Observation'
...
    ProblemDetail:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT token obtained from /api/auth/login'